+++ src/modules/backend/world/Map.h
@@ -127,7 +127,7 @@ public:
 	int npcCount() const;
 	int userCount() const;
 
-	int findFloor(const glm::vec3& pos, float maxDistanceY = (float)voxel::MAX_HEIGHT) const;
+	int findFloor(const glm::ivec3& pos, int maxDistanceY = voxel::MAX_HEIGHT) const;
 	glm::ivec3 randomPos() const;
 
 	const DBChunkPersisterPtr& chunkPersister();
